fix: warnings

+++ cuda/src/harris.cu
@@ -402,7 +402,8 @@ __global__ void best_points(char* harris_resp, point* points, char* values,
 }
 
 void harris(char* host_buffer, char* out_buffer, point* out_point,
-            int* nb_points, size_t width, size_t height, std::ptrdiff_t stride)
+            size_t* nb_points, size_t width, size_t height,
+            std::ptrdiff_t stride)
 {
     cudaError_t rc = cudaSuccess;
